backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
sentence-transformers==2.2.2
openai==1.3.7
faker==20.1.0

